Add nvidia Grace Architecture (#7158)
* Add Grace CPU architecture

* Apply suggestions from code review

Co-authored-by: Damien L-G <dalg24+github@gmail.com>
Co-authored-by: Daniel Arndt <arndtd@ornl.gov>

* Disable NEON simd when compiling with NVCC

* Fix clang format

* Test if compile flags exist before setting them

Only for CMake

* Add one more check for nvcc compiler

We can not use KOKKOS_ARCH_ARM_NEON on Grace with nvcc. How,ever one check was misisng and we included simd there but did not include the headers. With this last check we make sure to use scalar with nvcc.

* Update Kokkos_SIMD.hpp

Remove the check for nvc here and add it to cmake.

* Update kokkos_arch.cmake

Disable NEON on ARM if CUDA is enabled since nvc can not compile that.

* SIMD: fix clang-format

* Update cmake/kokkos_arch.cmake

Co-authored-by: Daniel Arndt <arndtd@ornl.gov>

* Update cmake/kokkos_arch.cmake

Co-authored-by: Damien L-G <dalg24+github@gmail.com>

* simd: no check for NVCC

* Remove unrelated line change

---------

Co-authored-by: Christian Trott <crtrott@sandia.gov>
Co-authored-by: Damien L-G <dalg24+github@gmail.com>
Co-authored-by: Daniel Arndt <arndtd@ornl.gov>
Co-authored-by: Christian Robert Trott <crtrott@grace1.eth>
Co-authored-by: Patrick Diehl <diehlpk@users.noreply.github.com>