Use absolute value of delta in conversion test for b16 (#7593)
* Use absolute value of delta in conversion test for half

Test for conversion to and from half checked that the delta between the expected
value and the result of the conversion were below some bound.
Since this delta was negative and no absolute value was used, tests used to pass
despite the result beeing out of bound.

* Replaces ASSERT_LT with ASSERT_NEAR for f16 and b16 conversion tests

Also adds details on how the bounds have been chosen for the tests

* Correct format with clang-format

* Apply suggestions from code review

Co-authored-by: Daniel Arndt <arndtd@ornl.gov>

---------

Co-authored-by: paul.gannay@cea.fr <gannayp@ruche02.cluster>
Co-authored-by: paul.gannay@cea.fr <gannayp@ruche01.cluster>
Co-authored-by: Paul Gannay <paul.gannay@cea.fr>
Co-authored-by: Daniel Arndt <arndtd@ornl.gov>