GCC minimum version for cxx 20 (#8130)
* GCC minimum version for cxx 20

Proposing to raise the minimum version of gcc to avoid
running into the issue of LWG-3446: https://cplusplus.github.io/LWG/issue3446
This issue is causing failures in Trilinos but not in Kokkos yet.
Note that the gcc versions currently affected by LWG-3446 are:
 - gcc < 10.4.0
 - gcc < 11.1.0
 - clang < 14

Please let me know if the proposed change of minimum gcc raises
issues with any important customer and/or platform.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Add warning message for GCC 11.0 when using C++20

Signed-off-by: tcclevenger <tccleve@sandia.gov>

* Fix typo

Co-authored-by: Daniel Arndt <arndtd@ornl.gov>

* Format

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>
Signed-off-by: tcclevenger <tccleve@sandia.gov>
Co-authored-by: tcclevenger <tccleve@sandia.gov>
Co-authored-by: Damien L-G <dalg24+github@gmail.com>
Co-authored-by: Daniel Arndt <arndtd@ornl.gov>